refactor: Separate markParsing into its own component
