+++ src/lib/ChatScreens/ChatBody.svelte
@@ -1,0 +1,154 @@
+<script lang="ts">
+    import { isEqual } from "lodash"
+    import { DBState } from 'src/ts/stores.svelte'
+    import { sleep } from "src/ts/util"
+    import { alertError } from "../../ts/alert"
+    import { ParseMarkdown, postTranslationParse, type CbsConditions, type simpleCharacterArgument } from "../../ts/parser.svelte"
+    import { getLLMCache, translateHTML } from "../../ts/translator/translator"
+    
+    interface Props {
+        character?: simpleCharacterArgument|string|null
+        firstMessage?: boolean
+        idx?: any
+        msgDisplay?: string
+        name?: string
+        role: string|null
+        translated: boolean
+        translating: boolean
+    }
+
+    let {
+        character = null,
+        idx = 0,
+        firstMessage = false,
+        msgDisplay,
+        role,
+        translated = $bindable(false),
+        translating = $bindable(false)
+    }: Props =  $props()
+
+    // svelte-ignore non_reactive_update
+    let lastParsed = ''
+    let lastCharArg:string|simpleCharacterArgument = null
+    let lastChatId = -10
+    let retranslate = false
+
+    function getCbsCondition(){
+        try{
+            const cbsConditions:CbsConditions = {
+                firstmsg: firstMessage ?? false,
+                chatRole: role,
+            }
+            return cbsConditions
+        }
+        catch(e){
+            return {
+                firstmsg: firstMessage ?? false,
+                chatRole: null,
+            }
+        }
+    }
+    
+    const markParsing = async (data: string, charArg?: string | simpleCharacterArgument, mode?: "normal" | "back", chatID?: number, translateText?:boolean, tries?:number) => {
+        let lastParsedQueue = ''
+        try {
+            if((!isEqual(lastCharArg, charArg)) || (chatID !== lastChatId)){
+                lastParsedQueue = ''
+                lastCharArg = charArg
+                lastChatId = chatID
+                translateText = false
+                try {
+                    if(DBState.db.autoTranslate){
+                        if(DBState.db.autoTranslateCachedOnly && DBState.db.translatorType === 'llm'){
+                            const cache = DBState.db.translateBeforeHTMLFormatting
+                            ? await getLLMCache(data)
+                            : !DBState.db.legacyTranslation
+                            ? await getLLMCache(await ParseMarkdown(data, charArg, 'pretranslate', chatID, getCbsCondition()))
+                            : await getLLMCache(await ParseMarkdown(data, charArg, mode, chatID, getCbsCondition()))
+                  
+                            translateText = cache !== null
+                        }
+                        else{
+                            translateText = true
+                        }
+                    }
+
+                    const lastTranslated = translated
+
+                    setTimeout(() => {
+                            translated = translateText
+                    }, 10)
+
+                    // State change of `translated` triggers markParsing again,
+                    // causing redundant translation attempts
+                    if (lastTranslated !== translateText) {
+                        return;
+                    }
+                } catch (error) {
+                    console.error(error)
+                }
+            }
+            if(translateText){
+                if (!retranslate && DBState.db.showTranslationLoading) {
+                    lastParsed = `<div class="flex justify-center items-center"><div class="animate-spin rounded-full h-8 w-8 border-b-2 border-textcolor"></div></div>`
+                }
+                
+                let doRetranslate = retranslate
+                retranslate = false
+                if(DBState.db.translatorType === 'llm' && DBState.db.translateBeforeHTMLFormatting){
+                    await sleep(100)
+                    translating = true
+                    data = await translateHTML(data, false, charArg, chatID, doRetranslate)
+                    translating = false
+                    const marked = await ParseMarkdown(data, charArg, mode, chatID, getCbsCondition())
+                    lastParsedQueue = marked
+                    lastCharArg = charArg
+                    return marked
+                }
+                else if(!DBState.db.legacyTranslation){
+                    const marked = await ParseMarkdown(data, charArg, 'pretranslate', chatID, getCbsCondition())
+                    translating = true
+                    const translated = await postTranslationParse(await translateHTML(marked, false, charArg, chatID, doRetranslate))
+                    translating = false
+                    lastParsedQueue = translated
+                    lastCharArg = charArg
+                    return translated
+                }
+                else{
+                    const marked = await ParseMarkdown(data, charArg, mode, chatID, getCbsCondition())
+                    translating = true
+                    const translated = await translateHTML(marked, false, charArg, chatID, doRetranslate)
+                    translating = false
+                    lastParsedQueue = translated
+                    lastCharArg = charArg
+                    return translated
+                }
+            }
+            else{
+                const marked = await ParseMarkdown(data, charArg, mode, chatID, getCbsCondition())
+                lastParsedQueue = marked
+                lastCharArg = charArg
+                return marked
+            }   
+        } catch (error) {
+            //retry
+            if(tries > 2){
+
+                alertError(`Error while parsing chat message: ${translateText}, ${error.message}, ${error.stack}`)
+                return data
+            }
+            return await markParsing(data, charArg, mode, chatID, translateText, (tries ?? 0) + 1)
+        }
+        finally{
+            lastParsed = lastParsedQueue
+        }
+    }
+
+    let markParsingResult = $derived.by(() => markParsing(msgDisplay, character, 'normal', idx, translated))
+</script>
+
+{#await markParsingResult}
+    {@html lastParsed}
+{:then md}
+    {@html md}
+{/await}
